requirements.txt
pyserial
Pillow
screeninfo